Update docs: LSCR registry configuration
While configuring LSCR, I noticed that the ENV variables configured
weren't working as expected/documented.

I took some guesses and figured out that the following variables
worked:

|Env|Usage|
|---|-----|
|`WUD_REGISTRY_LSCR_USERNAME`|GitHub username|
|`WUD_REGISTRY_LSCR_TOKEN`|GitHub PAT|

I updated the docs accordingly and adjusted the username `john@doe`
in the examples to `johndoe` to make it clearer that this is a
username - I was a bit confused with the `@` when I read the docs.